int types handling

+++ code/diff2d/lib/oned.hpp
@@ -27,8 +27,8 @@ namespace linalg {
     //codesnippet end
 
     // constructors
-    bordered_array_1d( size_t m,size_t n,int border );
-    bordered_array_1d( size_t m,size_t n,real *data )
+    bordered_array_1d( int64_t m,int64_t n,int border );
+    bordered_array_1d( int64_t m,int64_t n,real *data )
       : bordered_array_base<real>(m,n,data) {};
 
     // required functionality
@@ -44,18 +44,18 @@ namespace linalg {
      */
     
     //! convert linear ij interior to pair i,j in the global data
-    std::pair<size_t,size_t> split_i_j( size_t ij ) {
+    std::pair<int64_t,int64_t> split_i_j( int64_t ij ) {
       return std::make_pair( ij/this->n()+this->border(), ij%this->n()+this->border() );
     };
 
     //! m/n/border size of the allocated data
-    std::tuple<size_t,size_t,size_t> outer_sizes() { 
+    std::tuple<int64_t,int64_t,int64_t> outer_sizes() { 
       return std::make_tuple
         ( this->m()+2*this->border(),
           this->n()+2*this->border(),this->border() ); };
 
     //! m/n size of the domain, and the border
-    std::tuple<size_t,size_t,size_t> inner_sizes() { 
+    std::tuple<int64_t,int64_t,int64_t> inner_sizes() { 
       return std::make_tuple( this->m(),this->n(),this->border() ); };
 
     std::vector<real> internal_data();
